step2 (fix) Ladder 역할 위임
